:art: refactor codes
