+++ src/pages/games/tetris/composables/useI.ts
@@ -1,8 +1,9 @@
 import { useNodes } from './useNodes';
 import { usePositions } from './usePositions';
 import { NODE } from './@types';
+import type { UseTetrominoResult } from './@types';
 
-export function useI() {
+export function useI(): UseTetrominoResult {
   /**
    * ### Shape 0
    * ```
