+++ src/pages/games/tetris/composables/useL.ts
@@ -1,8 +1,9 @@
 import { useNodes } from './useNodes';
 import { usePositions } from './usePositions';
 import { NODE } from './@types';
+import type { UseTetrominoResult } from './@types';
 
-export function useL() {
+export function useL(): UseTetrominoResult {
   /**
    * ### Shape 0
    * ```
