+++ src/pages/games/tetris/composables/useS.ts
@@ -1,8 +1,9 @@
 import { useNodes } from './useNodes';
 import { usePositions } from './usePositions';
 import { NODE } from './@types';
+import type { UseTetrominoResult } from './@types';
 
-export function useS() {
+export function useS(): UseTetrominoResult {
   /**
    * ### Shape 0
    * ```
