+++ src/pages/games/tetris/composables/useScore.ts
@@ -1,0 +1,1 @@
+export function useScore() {}
