+++ src/pages/games/tetris/composables/useZ.ts
@@ -1,8 +1,9 @@
 import { useNodes } from './useNodes';
 import { usePositions } from './usePositions';
 import { NODE } from './@types';
+import type { UseTetrominoResult } from './@types';
 
-export function useZ() {
+export function useZ(): UseTetrominoResult {
   /**
    * ### Shape 0
    * ```
